feat: refactor function by exporting it
